Renamed itf/plugins/base/test to itf/plugins/base/utils
